反射调整

+++ Reflection/GUI/NumericFieldDraggerUtility_Internal.cs
@@ -4,7 +4,7 @@ namespace Emilia.Reflection.Editor
 {
     public static class NumericFieldDraggerUtility_Internal
     {
-        public static double CalculateIntDragSensitivity_Internal(long value)
+        public static long CalculateIntDragSensitivity_Internal(long value)
         {
             return NumericFieldDraggerUtility.CalculateIntDragSensitivity(value);
         }
